tracking failure message display
